feat: improved channel locking system (#18)

+++ init.sh
@@ -7,4 +7,9 @@ mkdir -p storage
 
 touch logs/join-leave.log
 echo "{}" > config/snippets.json
+
+if [ -e config/config.json ]; then
+    exit;
+fi
+
 cp sample-config.json config/config.json
